+++ app/src/auth/AppAuth.ts
@@ -1,6 +1,7 @@
-import { Authenticator, AuthPermissions, Role, type Strategy } from "auth";
-import type { PasswordStrategy } from "auth/authenticate/strategies";
 import type { DB } from "bknd";
+import * as AuthPermissions from "auth/auth-permissions";
+import type { AuthStrategy } from "auth/authenticate/strategies/Strategy";
+import type { PasswordStrategy } from "auth/authenticate/strategies/PasswordStrategy";
 import { $console, secureRandomString, transformObject } from "core/utils";
 import type { Entity, EntityManager } from "data";
 import { em, entity, enumm, type FieldSchema } from "data/prototype";
@@ -10,6 +11,8 @@ import { type AppAuthSchema, authConfigSchema, STRATEGIES } from "./auth-schema"
 import { AppUserPool } from "auth/AppUserPool";
 import type { AppEntity } from "core/config";
 import { usersFields } from "./auth-entities";
+import { Authenticator } from "./authenticate/Authenticator";
+import { Role } from "./authorize/Role";
 
 export type UserFieldSchema = FieldSchema<typeof AppAuth.usersFields>;
 declare module "bknd" {
@@ -88,7 +91,7 @@ export class AppAuth extends Module<AppAuthSchema> {
       this.ctx.guard.registerPermissions(AuthPermissions);
    }
 
-   isStrategyEnabled(strategy: Strategy | string) {
+   isStrategyEnabled(strategy: AuthStrategy | string) {
       const name = typeof strategy === "string" ? strategy : strategy.getName();
       // for now, password is always active
       if (name === "password") return true;
