+++ app/src/modules/Controller.ts
@@ -1,7 +1,6 @@
-import type { App } from "App";
+import type { App, SafeUser } from "bknd";
 import { type Context, type Env, Hono } from "hono";
 import * as middlewares from "modules/middlewares";
-import type { SafeUser } from "auth";
 import type { EntityManager } from "data";
 import { s } from "bknd/utils";
